% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{eq_read_data}
\alias{eq_read_data}
\title{Mastering Software Development in R Specialization Capstone Project
Data loading function
The function loads the raw NOAA data from a given file to a dataframe
Data loading function
The function loads the raw NOAA data from a given file to a dataframe}
\usage{
eq_read_data(filename)
}
\arguments{
\item{filename}{A name of the file}
}
\value{
dataset
}
\description{
Mastering Software Development in R Specialization Capstone Project
Data loading function
The function loads the raw NOAA data from a given file to a dataframe
Data loading function
The function loads the raw NOAA data from a given file to a dataframe
}
\examples{
\dontrun{
  sample_file <- eq_clean_data("signif.txt")
}

}
