% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{location_clean}
\alias{location_clean}
\title{Location cleaning function
The function takes NOAA data and cleans the LOCATION_NAME column
by stripping out the country name (including the colon) and
converting names to title case}
\usage{
location_clean(raw_data)
}
\arguments{
\item{raw_data}{The raw NOAA dataset}
}
\value{
Dataset with cleaned LOCATION_NAME column
}
\description{
Location cleaning function
The function takes NOAA data and cleans the LOCATION_NAME column
by stripping out the country name (including the colon) and
converting names to title case
}
\examples{
\dontrun{
  clean_data <- location_clean(raw_data)
}

}
